rag-backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-backend"
version = "0.1.0"
description = "RAG Backend for AI Textbook Application"
authors = [
    {name = "AI Textbook Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "qdrant-client>=1.7.1",
    "pydantic>=2.5.0,<3.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.3",
    "httpx>=0.25.2",
    "cohere>=4.36",
    "google-generativeai>=0.8.4",
    "mangum>=0.17.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.scripts]
start = "src.main:main"